requirements.txt
asgiref==3.6.0
boto3==1.20.9
botocore==1.23.54
certifi==2022.12.7
charset-normalizer==2.0.12
Django==4.2.1
django-ckeditor==6.1.0
django-js-asset==2.0.0
django-storages==1.13.2
idna==3.4
jmespath==0.10.0
python-dateutil==2.8.2
python-decouple==3.5
requests==2.26.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
